CodecTight.as:  Force screen repaint after blitting jpeg image.
Due to the two-deque system that the guy writing ActionScript used, it
was possible (and frequent) to get the following scenario:
- receive RFB message saying "Tight-encoded JPEG coming."
- begin decoding jpeg (asynchronously).
- while decoding, signal that the screen was updated.
- finish decoding jpeg.

OOPS!  If we don't tell Flash the screen was updated (by unlock()ing
it), Flash doesn't repaint the screen.  Throw in an extra repaint in
there once we're done the asynchronous jpeg decode, and it all works.